Update patch set 4
Patch Set 4:

Please consider this revised version.  I took the liberty of deleting the Dockerfiles that are no longer required for Jenkins-based continuous integration.  I do NOT plan to reorganize files into a subdirectory because the convention for Docker projects is that they have a Dockerfile at their root.

Patch-set: 4