Update patch set 1
Patch Set 1:

(4 comments)

I can offer comments about the CI -- basically just the Dockerfile.

I think the Dashboard is in a similar situation as this, in that it also needs to call the App Manager.  The dashboard project has a cached copy of the AppMgr swagger spec (just like here), but in the dashboard build process the client is generated on the fly, and no generated code is committed to Gerrit. Please consider if a similar approach would work here.  

So I would like to vote -1 bcos of committing generated coe, but I don't want to block your work at this late date, so I will give no vote at all.

I'm not qualified to review your Go code, sorry.

Patch-set: 1
Label: Verified=0