Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > remerge
> > (for triggering build of docker image with new name: o-ran-sc prefix)
> 
> I didnt get this. This request is already merged. What is the expected from us ?

This is a trick in this gerrit-jenkins system of LF.  Normally Jenkins build jobs are triggered by gerrit code changes.  But to be able to trigger jobs at any time, LF defined a number of magic words, which is you type in reply the corresponding Jenkins Job will be triggered.  remerge is the magic word for triggering merge job.  I need to trigger this job to rebuild the docker image.

Patch-set: 4